refactor(operator): address review nits (shared SAR helper, YAML style)
- Move the SubjectAccessReview helper canCreateDeployment(), plus a small
  operatorServiceAccountUser() helper, into OLMITBase so the OLM ITs share one
  copy instead of duplicating the method.
- Convert the secrets rule in rbac/namespaced/role.yaml from flow style to block
  style to match every other rule in the file.
